pyproject.toml
[build-system]
requires = ["setuptools>=80", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "q2m3"
dynamic = ["version"]
description = "Quantum-Classical Hybrid QM/MM Framework POC with QPE for Early Fault-Tolerant Quantum Computers"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "IHPC Quantum Computing Team"},
]
keywords = ["quantum computing", "quantum chemistry", "QM/MM", "QPE", "PySCF", "PennyLane"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyscf>=2.0.0",
    "pennylane>=0.44.0",
    "numpy>=2.0.0",
    "scipy>=1.7.0",
    "pyyaml>=6.0",
    "matplotlib>=3.5.0",
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pre-commit>=4.0.0",
    "commitizen>=4.10.0",
    "ipython>=8.0.0",
    "jupyter>=1.0.0",
    "notebook>=6.5.0",
]
docs = [
    "sphinx>=8.0.0",
    "furo>=2024.8.6",
    "myst-parser>=4.0.0",
    "linkify-it-py>=2.0.0",
    "sphinx-copybutton>=0.5.2",
    "sphinxcontrib-mermaid>=1.0.0",
]
gpu = [
    "pennylane-lightning[gpu]>=0.44.0",
    "cupy-cuda12x>=12.0.0",  # CUDA 12/13 compatible
    "jax[cuda12]>=0.4.0",    # JAX with CUDA support
]
catalyst = [
    "pennylane-catalyst>=0.14.0",
    "pennylane-lightning>=0.44.0",
]
solvation = [
    "q2m3[catalyst]",
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
]
viz = [
    "py3dmol>=2.0.0",
    "ase>=3.22.0",
    "nglview>=3.0.0",
]

[project.urls]
"Homepage" = "https://github.com/yjmaxpayne/q2m3"
"Documentation" = "https://github.com/yjmaxpayne/q2m3#readme"
"Bug Tracker" = "https://github.com/yjmaxpayne/q2m3/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.xyz"]

[tool.setuptools_scm]
fallback_version = "0.1.0.dev0"

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["examples/_archived"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B904",  # raise from
]

[tool.ruff.lint.per-file-ignores]
"examples/h2_qpe_validation.py" = ["E402"]
"examples/h2_mc_solvation.py" = ["E402"]
"examples/h3o_mc_solvation.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["q2m3"]
known-third-party = ["pyscf", "pennylane", "numpy", "scipy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "--verbose",
    "--cov=src/q2m3",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "solvation: marks tests requiring solvation dependencies (Catalyst + JAX)",
    "catalyst: marks tests requiring Catalyst (skip if not installed)",
    "gpu: marks tests requiring lightning.gpu (skip if not available)",
    "rdm: marks tests related to 1-RDM measurement functionality",
]

[tool.coverage.run]
source = ["src/q2m3"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.commitizen]
name = "cz_conventional_commits"
version_provider = "scm"
tag_format = "v$version"
version_scheme = "semver"
changelog_file = "CHANGELOG.md"
update_changelog_on_bump = false
changelog_incremental = true
